# docs/topics/gamification_openapi_2.self-contained.yaml
...
openapi: "3.0.3"
info:
  title: "gamification API"
  description: "gamification API"
  version: "1.0.0"
servers:
  - url: "https://api.tc-syd-02.insa-lyon.fr"
    description: "Production server"
paths:
  /api/session/available:
    get:
      summary: "Get all available sessions"
      description: "Get all available sessions - Requires authentication"
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/StudentSession"
                  - type: array
                    items:
                      $ref: "#/components/schemas/TeacherSession"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/session/all:
    get:
      summary: "Get all sessions for user"
      description: "Get all sessions for user. Results type differ if user is student or teacher/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    description: "Returns sessions with student-specific information that are either in progress or done"
                    items:
                      $ref: "#/components/schemas/StudentSession"
                  - type: array
                    description: "Returns sessions with teacher-specific information that are either scheduled, in progress or done"
                    items:
                      $ref: "#/components/schemas/TeacherSession"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/session/new:
    post:
      summary: "Creates a new session"
      description: "Creates a new session. Only available to teachers/admins. Returns a TeacherSession after creation - Requires authentication"
      security:
        - cookieAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Session 1"
                password:
                  type: string
                  example: "Password123"
                startDate:
                  type: string
                  format: date-time
                  example: "2024-01-01T00:00:00.000Z"
                endDate:
                  type: string
                  format: date-time
                  example: "2024-01-01T02:00:00.000Z"
                TP:
                  type: string
                  enum:
                    - "scrapping"
                    - "kafka"
                  example: "scrapping"
                indexGrades:
                  type: object
                  example: { "0": 5, "1": 10, "2": 20 }
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherSession"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Validation error"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/session/fetchTPs:
    get:
      summary: "Get all TPs strings"
      description: "Get all TPs strings. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
  /api/session/{id}:
    get:
      summary: "Get session by id"
      description: "Get session by id. Returns a StudentSession if user is student, TeacherSession if user is teacher/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          example: "65a7abe1ef63234c1129d7f9"
          required: true
          schema:
            type: string
          description: The session id
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StudentSession"
                  - $ref: "#/components/schemas/TeacherSession"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      summary: "Delete session by id"
      description: "Delete session by id. Only available to teachers/admin. Cannot delete sessions that have already started - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session deleted"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session already started"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/session/{id}/end:
    post:
      summary: "End session by id"
      description: "End session by id. Only available to teachers/admin. Cannot end sessions that have already ended - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherSession"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session already ended"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/session/{id}/join:
    post:
      summary: "Join session by id"
      description: "Join session by id. Only available to students. Cannot join sessions that have not started or that have ended - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: "#/components/schemas/StudentSession"
                  token:
                    type: string
                  progression:
                    $ref: "#/components/schemas/Progression"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session has not started yet"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/session/{id}/edit:
    post:
      summary: "POST api/session/{id}/edit"
      description: "Edit session by id. Only available to teachers/admin. For already started sessions, only the password and the endDate can be edited - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Session 1"
                password:
                  type: string
                  example: "Password123"
                startDate:
                  type: string
                  format: date-time
                  example: "2024-01-01T00:00:00.000Z"
                endDate:
                  type: string
                  format: date-time
                  example: "2024-01-01T02:00:00.000Z"
                TP:
                  type: string
                  enum:
                    - "scrapping"
                    - "kafka"
                  example: "scrapping"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherSession"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...
components:
  schemas:
    StudentSession:
      description: "A session with student-specific information"
      type: "object"
      properties:
        id:
          description: "The id of the session"
          type: "string"
        name:
          description: "The name of the session"
          type: "string"
        teachers:
          description: "A list of the teachers' user ids for this session"
          type: "array"
          items:
            type: "string"
        startDate:
          description: "The start date of the session"
          type: "string"
          format: "date-time"
        endDate:
          description: "The end date of the session"
          type: "string"
          format: "date-time"
        TP:
          description: "The type of the TP for this session"
          type: "string"
        indexGrades:
          description: "A map of level to grade (string to integer)"
          type: "object"
        status:
          description: "The status of the session"
          enum:
            - "inProgress"
            - "done"
          type: "string"
        joined:
          description: "Whether the current user has joined the session"
          type: "boolean"
    TeacherSession:
      description: "A session with teacher-specific information"
      type: "object"
      properties:
        id:
          description: "The id of the session"
          type: "string"
        name:
          description: "The name of the session"
          type: "string"
        teachers:
          description: "A list of the teachers' user ids for this session"
          type: "array"
          items:
            type: "string"
        students:
          description: "A list of the students' user ids for this session"
          type: "array"
          items:
            type: "string"
        startDate:
          description: "The start date of the session"
          type: "string"
          format: "date-time"
        endDate:
          description: "The end date of the session"
          type: "string"
          format: "date-time"
        TP:
          description: "The type of the TP for this session"
          type: "string"
        indexGrades:
          description: "A map of level to grade (string to integer)"
          type: "object"
        status:
          description: "The status of the session"
          enum:
            - "scheduled"
            - "inProgress"
            - "done"
          type: "string"
    Progression:
      description: "A progression"
      type: "object"
      properties:
        userId:
          description: "The student's user id"
          type: "string"
        sessionId:
          description: "The session's id"
          type: "string"
        level:
          description: "The level of the student in the TP"
          type: "integer"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid